重新划分模块
将随机数相关单独划分模块
