+++ ChanceCustom/replyBase.py
@@ -15,7 +15,6 @@ _________ ___________________ ____  __.
 
 import OlivOS
 import ChanceCustom
-import random
 import os
 import sys
 
@@ -286,67 +285,6 @@ def updateValFunTemp():
             return res
         return updateVal_f
     return updateValFun
-
-def RangeNumFunTemp():
-    def RangeNumFun(valDict):
-        def RangeNum_f(matched:'re.Match|dict'):
-            groupDict = getGroupDictInit(matched)
-            res = 0
-            resDict = {}
-            getCharRaw(resDict, 'X', '0', groupDict)
-            getCharRaw(resDict, 'Y', None, groupDict)
-            if resDict['Y'] == None:
-                if '-' in resDict['X']:
-                    try:
-                        tmp = resDict['X'].split('-')
-                        x = int(tmp[0])
-                        y = int(tmp[1])
-                    except:
-                        x = 0
-                        y = 0
-                else:
-                    x = 0
-                    y = 0
-            else:
-                getNumRegTatol(resDict, 'X', '0', groupDict, valDict)
-                getNumRegTatol(resDict, 'Y', '0', groupDict, valDict)
-                x = resDict['X']
-                y = resDict['Y']
-            res = random.randint(x,y)
-            return str(res)
-        return RangeNum_f
-    return RangeNumFun
-
-def PaddingRangeNumFunTemp():
-    def PaddingRangeNumFun(valDict):
-        def PaddingRangeNum_f(matched:'re.Match|dict'):
-            groupDict = getGroupDictInit(matched)
-            res = 0
-            resDict = {}
-            getCharRaw(resDict, 'X-Y', '0-0', groupDict)
-            try:
-                tmp = resDict['X-Y'].split('-')
-                x = int(tmp[0])
-                y = int(tmp[1])
-            except:
-                x = 0
-                y = 0
-            res = str(random.randint(x,y))
-            length = len(str(y))
-            res = res.zfill(length)
-            return res
-        return PaddingRangeNum_f
-    return PaddingRangeNumFun
-
-def RangeCharFunTemp():
-    def RangeCharFun(valDict):
-        def RangeChar_f(matched:'re.Match|dict'):
-            groupDict = getGroupDictInit(matched)
-            alphabet = 'AaBbCcDdEdFfGgHhIiJjKkLlMmNnOoPpQqRrSsTtUuVvWwXxYyZz0123456789'
-            res = random.choice(alphabet)
-            return res
-        return RangeChar_f
-    return RangeCharFun
 
 def RunDirectoryFunTemp():
     def RunDirectoryFun(valDict):
